Update README.md re: 32 bit code
I've deleted the Shredos i686 repository as it was outdated, duplicated content in ShredOS.x86_64 and a waste of time maintaining two separate repositories that were almost identical. Instead I will be adding instructions on how modify the ShredOS.x86_64 build so that it generates 32 bit code and .iso files for those that wish to build from source.