Various fixes to do with tftp/ftp transfers
1. Debug tftp transfers for restoring config files and outputing
pdfs and logs.

2. Added a ping status delay for the ftp/tftp that only proceeds
with launching nwipe if ftp/tftp servers have been configured on
the kernel command line. The delay has a 30 second timeout upon
which nwipe will launch. This fixes a problem where nwipe launches
before the ethernet hardware is active and a IP address had not
been obtained. This caused nwipe to not be able to read the
config files from the ftp/tftp server.

3. Removed 4 second countdown on restarting nwipe.