feat(configs): adapt configurations for new changes
The default configuration "shredos_defconfig" was amended to produce both the existing USB image (which allows writing PDFs to the USB) and the new hybrid ISO (which now allows BIOS+UEFI booting, as well as flashing to USB and burning to CD/DVD-ROM) in one build.

Additional configurations "shredos_img_defconfig" and "shredos_iso_defconfig" were added which only target either the existing USB image or the new hybrid ISO, for full flexibility.

Signed-off-by: desertwitch <24509509+desertwitch@users.noreply.github.com>