feat(configs,shredos): add aio iso configuration
This adds a configuration for a jack-of-all-trades ISO, which allows BIOS and UEFI booting, being written to CD/DVD-ROM and USB drives, as well as having a writeable partition when being used from an USB drive. It basically combines the functionality of all prior release versions into one single ISO (per architecture).

Signed-off-by: desertwitch <24509509+desertwitch@users.noreply.github.com>