fix(configs): revert partition size and post processing for EFI
The EFI partition for hybrid ISOs does not reliably show on all Windows systems, so we cannot use our "hack" of writing the configurations/PDFs to it. Reverted instead to a sane-sized EFI partition which is used just for booting. Added a note to the README that some tools such as Rufus support an ISO-mode to add a writeable partition at burning-time.

Signed-off-by: desertwitch <24509509+desertwitch@users.noreply.github.com>