package/unzip: fix conflicting gmtime/localtime prototypes with GCC 15
GCC 15 enforces stricter prototype checks for standard library
functions. The unzip 6.0 sources declare their own prototypes for
gmtime() and localtime() in unix/unxcfg.h without parameters, which
conflicts with the libc-provided declarations that expect a `const time_t *`.

This results in the following build failures:

    error: conflicting types for 'gmtime'
    error: conflicting types for 'localtime'
    error: too many arguments to function 'localtime'
    error: too many arguments to function 'gmtime'

The custom prototypes in unxcfg.h are unnecessary on modern systems,
and removing them restores correct type checking and fixes the build
with GCC 15.

This patch deletes the outdated declarations:

    struct tm *gmtime(), *localtime();

Tested-by: Fabian Druschke fdruschke@outlook.com
Signed-off-by: Fabian Druschke fdruschke@outlook.com